edit file panel

+++ ImView/ischodn_dannie.cpp
@@ -75,7 +75,7 @@ void ischodn_dannie::on_pushButton_clicked()
     base.kpd_nom = ui->tableWidget->item(5,1)->text().toDouble();
     base.muk = ui->tableWidget->item(6,1)->text().toDouble();
     base.n_0 = ui->tableWidget->item(7,1)->text().toDouble();
-    base.ki = ui->tableWidget->item(8,1)->text().toDouble();
+    base.ki = ui->tableWidget->item(8,1)->text();
 
     if(ui->tableWidget->item(0,1)->text().isEmpty()||
             ui->tableWidget->item(1,1)->text().isEmpty()||
